+++ crates/networking/p2p/discv4/server.rs
@@ -39,8 +39,9 @@ const REVALIDATION_INTERVAL: Duration = Duration::from_secs(12 * 60 * 60); // 12
 /// The initial interval between peer lookups, until the number of peers reaches
 /// [target_peers](DiscoverySideCarState::target_peers), or the number of
 /// contacts reaches [target_contacts](DiscoverySideCarState::target_contacts).
-const INITIAL_LOOKUP_INTERVAL: Duration = Duration::from_secs(5);
-const LOOKUP_INTERVAL: Duration = Duration::from_secs(5 * 60); // 5 minutes
+pub const INITIAL_LOOKUP_INTERVAL: Duration = Duration::from_millis(100); // 10 per second
+pub const LOOKUP_INTERVAL: Duration = Duration::from_millis(600); // 100 per minute
+const CHANGE_FIND_NODE_MESSAGE_INTERVAL: Duration = Duration::from_secs(5);
 const PRUNE_INTERVAL: Duration = Duration::from_secs(5);
 
 #[derive(Debug, thiserror::Error)]
@@ -65,6 +66,7 @@ pub enum InMessage {
     Revalidate,
     Lookup,
     Prune,
+    ChangeFindNodeMessage,
     Shutdown,
 }
 
@@ -80,6 +82,9 @@ pub struct DiscoveryServer {
     signer: SecretKey,
     udp_socket: Arc<UdpSocket>,
     peer_table: PeerTable,
+    /// The last `FindNode` message sent, cached due to message
+    /// signatures being expensive.
+    find_node_message: BytesMut,
 }
 
 impl DiscoveryServer {
@@ -100,6 +105,7 @@ impl DiscoveryServer {
             signer,
             udp_socket,
             peer_table: peer_table.clone(),
+            find_node_message: Self::random_message(&signer),
         };
 
         info!(count = bootnodes.len(), "Adding bootnodes");
@@ -111,7 +117,7 @@ impl DiscoveryServer {
             .new_contacts(bootnodes, local_node.node_id())
             .await?;
 
-        discovery_server.start_on_thread();
+        discovery_server.start();
         Ok(())
     }
 
@@ -202,6 +208,18 @@ impl DiscoveryServer {
         Ok(())
     }
 
+    /// Generate and store a FindNodeMessage with a random key. We then send the same message on Disovery lookup.
+    /// We change this message every CHANGE_FIND_NODE_MESSAGE_INTERVAL.
+    fn random_message(signer: &SecretKey) -> BytesMut {
+        let expiration: u64 = get_msg_expiration_from_seconds(EXPIRATION_SECONDS);
+        let random_priv_key = SecretKey::new(&mut OsRng);
+        let random_pub_key = public_key_from_signing_key(&random_priv_key);
+        let msg = Message::FindNode(FindNodeMessage::new(random_pub_key, expiration));
+        let mut buf = BytesMut::new();
+        msg.encode_with_header(&mut buf, signer);
+        buf
+    }
+
     async fn revalidate(&mut self) -> Result<(), DiscoveryServerError> {
         for contact in self
             .peer_table
@@ -214,17 +232,9 @@ impl DiscoveryServer {
     }
 
     async fn lookup(&mut self) -> Result<(), DiscoveryServerError> {
-        // Sending the same FindNode message to all contacts to optimize message creation and encoding
-        let expiration: u64 = get_msg_expiration_from_seconds(EXPIRATION_SECONDS);
-        let random_priv_key = SecretKey::new(&mut OsRng);
-        let random_pub_key = public_key_from_signing_key(&random_priv_key);
-        let msg = Message::FindNode(FindNodeMessage::new(random_pub_key, expiration));
-        let mut buf = BytesMut::new();
-        msg.encode_with_header(&mut buf, &self.signer);
-
-        for contact in self.peer_table.get_contacts_for_lookup().await? {
-            if self.udp_socket.send_to(&buf, &contact.node.udp_addr()).await.inspect_err(
-                |e| error!(sending = ?msg, addr = ?&contact.node.udp_addr(), err=?e, "Error sending message"),
+        if let Some(contact) = self.peer_table.get_contact_for_lookup().await? {
+            if self.udp_socket.send_to(&self.find_node_message, &contact.node.udp_addr()).await.inspect_err(
+                |e| error!(sending = "FindNode", addr = ?&contact.node.udp_addr(), err=?e, "Error sending message"),
             ).is_err() {
                 self.peer_table
                     .set_disposable(&contact.node.node_id())
@@ -454,7 +464,7 @@ impl DiscoveryServer {
                 debug!(received = message_type, to = %format!("{sender_public_key:#x}"), "IP address mismatch, skipping");
                 Err(DiscoveryServerError::InvalidContact)
             }
-            PeerTableOutMessage::ValidContact(contact) => Ok(contact),
+            PeerTableOutMessage::Contact(contact) => Ok(contact),
             _ => unreachable!(),
         }
     }
@@ -505,8 +515,12 @@ impl GenServer for DiscoveryServer {
             InMessage::Revalidate,
         );
         send_interval(PRUNE_INTERVAL, handle.clone(), InMessage::Prune);
+        send_interval(
+            CHANGE_FIND_NODE_MESSAGE_INTERVAL,
+            handle.clone(),
+            InMessage::ChangeFindNodeMessage,
+        );
         let _ = handle.clone().cast(InMessage::Lookup).await;
-
         send_message_on(handle.clone(), tokio::signal::ctrl_c(), InMessage::Shutdown);
 
         Ok(Success(self))
@@ -547,6 +561,9 @@ impl GenServer for DiscoveryServer {
                     .prune()
                     .await
                     .inspect_err(|e| error!(err=?e, "Error Pruning peer table"));
+            }
+            Self::CastMsg::ChangeFindNodeMessage => {
+                self.find_node_message = Self::random_message(&self.signer);
             }
             Self::CastMsg::Shutdown => return CastResponse::Stop,
         }