+++ crates/networking/p2p/rlpx/initiator.rs
@@ -1,5 +1,10 @@
 use crate::{
-    discv4::peer_table::PeerTableError, metrics::METRICS, network::P2PContext,
+    discv4::{
+        peer_table::PeerTableError,
+        server::{INITIAL_LOOKUP_INTERVAL, LOOKUP_INTERVAL},
+    },
+    metrics::METRICS,
+    network::P2PContext,
     rlpx::connection::server::PeerConnection,
 };
 use spawned_concurrency::{
@@ -26,40 +31,32 @@ pub struct RLPxInitiator {
 
     /// The target number of RLPx connections to reach.
     target_peers: u64,
-    /// The rate at which to try new connections.
-    new_connections_per_lookup: usize,
 }
 
 impl RLPxInitiator {
     pub fn new(context: P2PContext) -> Self {
         Self {
             context,
-            initial_lookup_interval: Duration::from_secs(3),
-            lookup_interval: Duration::from_secs(5 * 60),
+            // We use the same lookup intervals as Discovery to try to get both process to check at the same rate
+            initial_lookup_interval: INITIAL_LOOKUP_INTERVAL,
+            lookup_interval: LOOKUP_INTERVAL,
             target_peers: 50,
-            new_connections_per_lookup: 5000,
         }
     }
 
     pub async fn spawn(context: P2PContext) {
         info!("Starting RLPx Initiator");
         let state = RLPxInitiator::new(context);
-        let mut server = RLPxInitiator::start_on_thread(state.clone());
-        let _ = server.cast(InMessage::LookForPeers).await;
+        let mut server = RLPxInitiator::start(state.clone());
+        let _ = server.cast(InMessage::LookForPeer).await;
     }
 
-    async fn look_for_peers(&mut self) -> Result<(), RLPxInitiatorError> {
-        debug!("Looking for peers");
+    async fn look_for_peer(&mut self) -> Result<(), RLPxInitiatorError> {
         if !self.context.table.target_peers_reached().await? {
-            let contacts = self
-                .context
-                .table
-                .get_contacts_to_initiate(self.new_connections_per_lookup)
-                .await?;
-            for contact in contacts {
+            if let Some(contact) = self.context.table.get_contact_to_initiate().await? {
                 PeerConnection::spawn_as_initiator(self.context.clone(), &contact.node).await;
                 METRICS.record_new_rlpx_conn_attempt().await;
-            }
+            };
         } else {
             debug!("Target peer connections reached, no need to initiate new connections.");
         }
@@ -80,7 +77,7 @@ impl RLPxInitiator {
 
 #[derive(Debug, Clone)]
 pub enum InMessage {
-    LookForPeers,
+    LookForPeer,
     Shutdown,
 }
 
@@ -106,17 +103,16 @@ impl GenServer for RLPxInitiator {
         handle: &GenServerHandle<Self>,
     ) -> CastResponse {
         match message {
-            Self::CastMsg::LookForPeers => {
-                debug!(received = "Look for peers");
+            Self::CastMsg::LookForPeer => {
                 let _ = self
-                    .look_for_peers()
+                    .look_for_peer()
                     .await
                     .inspect_err(|e| error!(err=?e, "Error looking for peers"));
 
                 send_after(
                     self.get_lookup_interval().await,
                     handle.clone(),
-                    Self::CastMsg::LookForPeers,
+                    Self::CastMsg::LookForPeer,
                 );
 
                 CastResponse::NoReply